absolute route import and module path aliases config
