+++ components/app/apply-pagination.js
@@ -1,0 +1,2 @@
+// TODO : refactor this to be global
+export const applyPagination = (tabs, page, rowsPerPage) => tabs.slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage);
