+++ hooks/use-auth.js
@@ -1,4 +1,4 @@
 import { useContext } from "react";
-import { AuthContext } from "../contexts/jwt-context";
+import { AuthContext } from "@/contexts/jwt-context";
 
 export const useAuth = () => useContext(AuthContext);
