+++ jsconfig.json
@@ -1,0 +1,12 @@
+{
+    "compilerOptions": {
+        "baseUrl": ".",
+        "paths": {
+            "@/components/*":["components/*"],
+            "@/contexts/*":["contexts/*"],
+            "@/hooks/*":["hooks/*"],
+            "@/lib/*":["lib/*"],
+            "@/faker/*":["faker/*"]
+        }
+    }
+}
